pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Data versioning for cybulde project using DVC"
authors = ["mdop297 <lenhatminh297@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
hydra-core = "~=1.3.2"
pydantic = "~=2.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "~=8.3.3"
black = "~=24.8.0"
isort = "~=5.13.2"
flake8 = "~=7.1.1"
mypy = "~=1.11.2"
jupyterlab = "~=4.2.5"

[tool.poetry.group.docs.dependencies]
sphinx = "~=8.0.2"

[tool.poetry.group.docs]
optional = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
indent = 4
lines_between_types = 1
known_first_party = ["src"]
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.mypy]
python_version = "3.11"
show_error_codes = true
ignore_missing_imports = true
install_types = true
non_interactive = true
follow_imports = "silent"
no_strict_optional = true

warn_redundant_casts = false
warn_unused_ignores = true
warn_unused_configs = true
warn_return_any = true
warn_no_return = true
warn_incomplete_stub = true

disallow_subclassing_any = true

disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
check_untyped_defs = true
strict_optional = true

verbosity = 0
